Optimized mem_set

+++ src/kernel/memory.h
@@ -2,7 +2,7 @@
 
 #include "types.h"
 
-void mem_set(void *ptr, unsigned char value, int size);
+void mem_set(void *ptr, unsigned char value, size_t size);
 void mem_copy(void *dst, const void *src, size_t size);
 
 void *mem_alloc(size_t size);
